backend/requirements.txt
flask
flask-cors
transformers
torch
torchvision
Pillow
requests
